+++ tests/typedb/bench/all.bench.ts
@@ -1,0 +1,4 @@
+import { allBench } from '../../unit/bench/bench';
+import { init } from '../helpers/init';
+
+allBench(init);