+++ tests/unit/bench/bench.ts
@@ -1,29 +1,28 @@
 import { v4 as uuidv4 } from 'uuid';
 
-import { deepRemoveMetaData, deepSort, expectArraysInObjectToContainSameElements } from '../../helpers/matchers';
-import type { typesSchema } from '../../mocks/generatedSchema';
-import type { TypeGen } from '../../../src/types/typeGen';
-import type { WithBormMetadata } from '../../../src/index';
-import type { UserType } from '../../types/testTypes';
 import { createTest } from '../../helpers/createTest';
-import { expect, it } from 'vitest';
+import { bench, expect } from 'vitest';
+import { deepRemoveMetaData, deepSort } from '../../helpers/matchers';
+import type { WithBormMetadata, TypeGen } from '../../../src';
+import type { typesSchema } from '../../mocks/generatedSchema';
+import type { UserType } from '../../types/testTypes';
 
-export const testQuery = createTest('Query', (ctx) => {
-	it('v1[validation] - $entity missing', async () => {
+export const allBench = createTest('Bench', (ctx) => {
+	bench('v1[validation] - $entity missing', async () => {
 		// @ts-expect-error - $entity is missing
 		await expect(ctx.query({})).rejects.toThrow();
 	});
 
-	it('v2[validation] - $entity not in schema', async () => {
+	bench('v2[validation] - $entity not in schema', async () => {
 		await expect(ctx.query({ $entity: 'fakeEntity' })).rejects.toThrow();
 	});
 
-	it('v3[validation] - $id not existing', async () => {
+	bench('v3[validation] - $id not existing', async () => {
 		const res = await ctx.query({ $entity: 'User', $id: 'nonExisting' });
 		await expect(res).toBeNull();
 	});
 
-	it('e1[entity] - basic and direct link to relation', async () => {
+	bench('e1[entity] - basic and direct link to relation', async () => {
 		const query = { $entity: 'User' };
 		const expectedRes = [
 			{
@@ -103,7 +102,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		expect(deepSort(res, 'id')).toEqual(expectedRes);
 	});
 
-	it('e1.b[entity] - basic and direct link to relation sub entity', async () => {
+	bench('e1.b[entity] - basic and direct link to relation sub entity', async () => {
 		const query = { $entity: 'God' };
 		const expectedRes = [
 			{
@@ -123,7 +122,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		expect(deepSort(res, 'id')).toEqual(expectedRes);
 	});
 
-	it('e2[entity] - filter by single $id', async () => {
+	bench('e2[entity] - filter by single $id', async () => {
 		const query = { $entity: 'User', $id: 'user1' };
 		const expectedRes = {
 			// '$entity': 'User',
@@ -151,7 +150,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		// expect(res['user-tags']).toHaveLength(expectedRes['user-tags'].length);
 	});
 
-	it('e3[entity, nested] - direct link to relation, query nested ', async () => {
+	bench('e3[entity, nested] - direct link to relation, query nested ', async () => {
 		const query = { $entity: 'User', $fields: ['id', { $path: 'user-tags' }] };
 		const expectedRes = [
 			{
@@ -256,7 +255,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		expect(deepSort(resWithoutMetadata, 'id')).toEqual(deepRemoveMetaData(expectedRes));
 	});
 
-	it('opt1[options, noMetadata', async () => {
+	bench('opt1[options, noMetadata', async () => {
 		const query = { $entity: 'User', $id: 'user1' };
 		const expectedRes = {
 			'name': 'Antoine',
@@ -280,7 +279,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		expect(res['user-tags']).toHaveLength(expectedRes['user-tags'].length);
 	});
 
-	it('TODO:opt2[options, debugger', async () => {
+	bench('TODO:opt2[options, debugger', async () => {
 		const query = { $entity: 'User', $id: 'user1' };
 		const expectedRes = {
 			'$id': 'user1',
@@ -338,7 +337,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		expect(res['user-tags']).toHaveLength(expectedRes['user-tags'].length);
 	});
 
-	it('opt3a[options, returnNull] - empty fields option in entity', async () => {
+	bench('opt3a[options, returnNull] - empty fields option in entity', async () => {
 		const query = {
 			$entity: 'User',
 			$id: 'user4',
@@ -358,7 +357,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		expect(deepSort(res, 'id')).toEqual(expectedRes);
 	});
 
-	it('opt3b[options, returnNull] - empty fields option in entity, dont return explicit', async () => {
+	bench('opt3b[options, returnNull] - empty fields option in entity, dont return explicit', async () => {
 		const query = {
 			$entity: 'User',
 			$id: 'user4',
@@ -377,7 +376,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		expect(deepSort(res, 'id')).toEqual(expectedRes);
 	});
 
-	it('r1[relation] - basic', async () => {
+	bench('r1[relation] - basic', async () => {
 		const query = { $relation: 'User-Accounts' };
 		const expectedRes = [
 			{
@@ -440,7 +439,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		);
 	});
 
-	it('r2[relation] - filtered fields', async () => {
+	bench('r2[relation] - filtered fields', async () => {
 		const query = { $relation: 'User-Accounts', $fields: ['user'] };
 		const expectedRes = [
 			{
@@ -486,7 +485,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		);
 	});
 
-	it('r3[relation, nested] - nested entity', async () => {
+	bench('r3[relation, nested] - nested entity', async () => {
 		const query = {
 			$relation: 'User-Accounts',
 			$fields: ['id', { $path: 'user', $fields: ['name'] }],
@@ -564,7 +563,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		expect(deepSort(resWithoutMetadata, 'id')).toEqual(deepRemoveMetaData(expectedRes));
 	});
 
-	it('r4[relation, nested, direct] - nested relation direct on relation', async () => {
+	bench('r4[relation, nested, direct] - nested relation direct on relation', async () => {
 		const query = {
 			$relation: 'UserTag',
 			$fields: [
@@ -623,7 +622,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		expect(deepSort(resWithoutMetadata, 'id')).toEqual(deepRemoveMetaData(expectedRes));
 	});
 
-	it('r5[relation nested] - that has both role, and linkfield pointing to same role', async () => {
+	bench('r5[relation nested] - that has both role, and linkfield pointing to same role', async () => {
 		const query = {
 			$entity: 'Color',
 			$fields: ['id', 'user-tags', 'group'],
@@ -658,7 +657,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		expect(deepSort(resWithoutMetadata, 'id')).toEqual(deepRemoveMetaData(expectedRes));
 	});
 
-	it('r6[relation nested] - relation connected to relation and a tunneled relation', async () => {
+	bench('r6[relation nested] - relation connected to relation and a tunneled relation', async () => {
 		const query = {
 			$relation: 'UserTag',
 		};
@@ -710,7 +709,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		expect(deepSort(resWithoutMetadata, 'id')).toEqual(deepRemoveMetaData(expectedRes));
 	});
 
-	it('r7[relation, nested, direct] - nested on nested', async () => {
+	bench('r7[relation, nested, direct] - nested on nested', async () => {
 		const query = {
 			$relation: 'UserTag',
 			$fields: [
@@ -844,7 +843,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		expect(deepSort(resWithoutMetadata, 'id')).toEqual(deepRemoveMetaData(expectedRes));
 	});
 
-	it('r8[relation, nested, deep] - deep nested', async () => {
+	bench('r8[relation, nested, deep] - deep nested', async () => {
 		const query = {
 			$entity: 'Space',
 			$id: 'space-2',
@@ -906,7 +905,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		expect(deepSort(resWithoutMetadata, 'id')).toEqual(deepRemoveMetaData(expectedRes));
 	});
 
-	it('r9[relation, nested, ids]', async () => {
+	bench('r9[relation, nested, ids]', async () => {
 		const query = {
 			$relation: 'UserTagGroup',
 			$id: 'utg-1',
@@ -925,7 +924,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		});
 	});
 
-	it('ef1[entity] - $id single', async () => {
+	bench('ef1[entity] - $id single', async () => {
 		const wrongRes = await ctx.query({ $entity: 'User', $id: uuidv4() });
 		expect(wrongRes).toEqual(null);
 		const validRes = await ctx.query({
@@ -936,7 +935,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		expect(validRes).toEqual({ $thing: 'User', $thingType: 'entity', $id: 'user1', id: 'user1' });
 	});
 
-	it('ef2[entity] - $id multiple', async () => {
+	bench('ef2[entity] - $id multiple', async () => {
 		const res = await ctx.query({
 			$entity: 'User',
 			$id: ['user1', 'user2'],
@@ -951,7 +950,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		]);
 	});
 
-	it('ef3[entity] - $fields single', async () => {
+	bench('ef3[entity] - $fields single', async () => {
 		const res = await ctx.query({ $entity: 'User', $fields: ['id'] });
 		expect(res).toBeDefined();
 		expect(res).not.toBeInstanceOf(String);
@@ -977,7 +976,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		]);
 	});
 
-	it('ef4[entity] - $fields multiple', async () => {
+	bench('ef4[entity] - $fields multiple', async () => {
 		const res = await ctx.query({
 			$entity: 'User',
 			$id: 'user1',
@@ -992,7 +991,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		});
 	});
 
-	it('ef5[entity,filter] - $filter single', async () => {
+	bench('ef5[entity,filter] - $filter single', async () => {
 		const res = await ctx.query({
 			$entity: 'User',
 			$filter: { name: 'Antoine' },
@@ -1002,7 +1001,8 @@ export const testQuery = createTest('Query', (ctx) => {
 		expect(res).toEqual([{ $thing: 'User', $thingType: 'entity', $id: 'user1', name: 'Antoine' }]);
 	});
 
-	it('ef6[entity,filter,id] - $filter by id in filter', async () => {
+	// can also be the id field!
+	bench('ef6[entity,filter] - $filter by unique field', async () => {
 		const res = await ctx.query({
 			$entity: 'User',
 			$filter: { id: 'user1' },
@@ -1011,10 +1011,10 @@ export const testQuery = createTest('Query', (ctx) => {
 		expect(res).toEqual({ $thing: 'User', $thingType: 'entity', $id: 'user1', name: 'Antoine' });
 	});
 
-	it('ef7[entity,unique] - $filter by unique field', async () => {
+	bench('ef7[entity,unique] - $filter unique field', async () => {
 		const res = await ctx.query({
 			$entity: 'User',
-			$filter: { email: 'antoine@test.com' },
+			$filter: { id: 'user1' }, // not $id, just being used as a regular field
 			$fields: ['name', 'email'],
 		});
 		// and now its not an array again, we used at least one property in the filter that is either the single key specified in idFields: ['id'] or has a validations.unique:true
@@ -1027,7 +1027,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		});
 	});
 
-	it('n1[nested] Only ids', async () => {
+	bench('n1[nested] Only ids', async () => {
 		const res = await ctx.query({
 			$entity: 'User',
 			$id: 'user1',
@@ -1045,7 +1045,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		});
 	});
 
-	it('n2[nested] First level all fields', async () => {
+	bench('n2[nested] First level all fields', async () => {
 		const query = {
 			$entity: 'User',
 			$id: 'user1',
@@ -1121,7 +1121,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		});
 	});
 
-	it('n3[nested, $fields] First level filtered fields', async () => {
+	bench('n3[nested, $fields] First level filtered fields', async () => {
 		const res = await ctx.query({
 			$entity: 'User',
 			$id: 'user1',
@@ -1141,7 +1141,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		});
 	});
 
-	it('n4a[nested, $id] Local filter on nested, by id', async () => {
+	bench('n4a[nested, $id] Local filter on nested, by id', async () => {
 		const res = await ctx.query({
 			$entity: 'User',
 			$id: ['user1', 'user2', 'user3'],
@@ -1186,7 +1186,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		]);
 	});
 
-	it('n4b[nested, $id] Local filter on nested depth two, by id', async () => {
+	bench('n4b[nested, $id] Local filter on nested depth two, by id', async () => {
 		const res = await ctx.query({
 			$entity: 'User',
 			$id: 'user1',
@@ -1218,7 +1218,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		});
 	});
 
-	it('nf1[nested, $filters] Local filter on nested, single id', async () => {
+	bench('nf1[nested, $filters] Local filter on nested, single id', async () => {
 		const query = {
 			$entity: 'User',
 			$id: 'user1',
@@ -1247,7 +1247,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		});
 	});
 
-	it('nf2[nested, $filters] Local filter on nested, by field, multiple sources, some are empty', async () => {
+	bench('nf2[nested, $filters] Local filter on nested, by field, multiple sources, some are empty', async () => {
 		const res = await ctx.query({
 			$entity: 'User',
 			$id: ['user1', 'user2', 'user3'],
@@ -1290,7 +1290,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		]);
 	});
 
-	it('nf3[nested, $filters] Local filter on nested, by link field, multiple sources', async () => {
+	bench('nf3[nested, $filters] Local filter on nested, by link field, multiple sources', async () => {
 		const res = await ctx.query({
 			$entity: 'Space',
 			$fields: [
@@ -1349,7 +1349,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		]);
 	});
 
-	it('nf4[nested, $filters] Local filter on nested, by link field, multiple sources', async () => {
+	bench('nf4[nested, $filters] Local filter on nested, by link field, multiple sources', async () => {
 		const res = await ctx.query({
 			$relation: 'UserTag',
 			$fields: [
@@ -1426,11 +1426,11 @@ export const testQuery = createTest('Query', (ctx) => {
 		]);
 	});
 
-	it('TODO:nf2a[nested, $filters] Nested filter for array of ids', async () => {
+	bench('TODO:nf2a[nested, $filters] Nested filter for array of ids', async () => {
 		expect(true).toEqual(false);
 	});
 
-	it('lf[$filter] Filter by a link field with cardinality ONE', async () => {
+	bench('lf[$filter] Filter by a link field with cardinality ONE', async () => {
 		const res = await ctx.query(
 			{
 				$relation: 'User-Accounts',
@@ -1442,7 +1442,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		expect(deepSort(res, 'id')).toMatchObject([{ id: 'ua1-1' }, { id: 'ua1-2' }, { id: 'ua1-3' }]);
 	});
 
-	it('lf[$filter] Filter out by a link field with cardinality ONE', async () => {
+	bench('lf[$filter] Filter out by a link field with cardinality ONE', async () => {
 		const res = await ctx.query(
 			{
 				$relation: 'User-Accounts',
@@ -1456,7 +1456,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		expect(deepSort(res, 'id')).toMatchObject([{ id: 'ua2-1' }, { id: 'ua3-1' }]);
 	});
 
-	it('lf[$filter] Filter by a link field with cardinality MANY', async () => {
+	bench('lf[$filter] Filter by a link field with cardinality MANY', async () => {
 		const res = await ctx.query(
 			{
 				$entity: 'User',
@@ -1468,7 +1468,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		expect(deepSort(res, 'id')).toMatchObject([{ id: 'user1' }, { id: 'user5' }]);
 	});
 
-	it('slo1[$sort, $limit, $offset] root', async () => {
+	bench('slo1[$sort, $limit, $offset] root', async () => {
 		const res = await ctx.query(
 			{
 				$entity: 'Account',
@@ -1488,7 +1488,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		]);
 	});
 
-	it('slo1[$sort, $limit, $offset] sub level', async () => {
+	bench('slo1[$sort, $limit, $offset] sub level', async () => {
 		const res = await ctx.query(
 			{
 				$entity: 'User',
@@ -1516,7 +1516,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		});
 	});
 
-	it('slo1[$sort, $limit, $offset] with an empty attribute', async () => {
+	bench('slo1[$sort, $limit, $offset] with an empty attribute', async () => {
 		const res = await ctx.query(
 			{
 				$entity: 'User',
@@ -1556,7 +1556,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		]);
 	});
 
-	it('i1[inherired, attributes] Entity with inherited attributes', async () => {
+	bench('i1[inherired, attributes] Entity with inherited attributes', async () => {
 		const res = await ctx.query({ $entity: 'God', $id: 'god1' }, { noMetadata: true });
 		expect(res).toEqual({
 			id: 'god1',
@@ -1567,7 +1567,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		});
 	});
 
-	it('s1[self] Relation playing a a role defined by itself', async () => {
+	bench('s1[self] Relation playing a a role defined by itself', async () => {
 		const res = await ctx.query({ $relation: 'Self' }, { noMetadata: true });
 		expect(deepSort(res, 'id')).toEqual([
 			{ id: 'self1', owned: ['self2'], space: 'space-2' },
@@ -1577,7 +1577,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		]);
 	});
 
-	it('ex1[extends] Query where an object plays 3 different roles because it extends 2 types', async () => {
+	bench('ex1[extends] Query where an object plays 3 different roles because it extends 2 types', async () => {
 		/// note: fixed with an ugly workaround (getEntityName() in parseTQL.ts)
 
 		const res = await ctx.query({ $entity: 'Space', $id: 'space-2' }, { noMetadata: true });
@@ -1593,7 +1593,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		});
 	});
 
-	it('ex2[extends] Query of the parent', async () => {
+	bench('ex2[extends] Query of the parent', async () => {
 		/// note: fixed with an ugly workaround (getEntityName() in parseTQL.ts)
 		const res = await ctx.query({ $entity: 'Space', $id: 'space-2', $fields: ['objects'] }, { noMetadata: true });
 		expect(deepSort(res, 'id')).toEqual({
@@ -1601,7 +1601,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		});
 	});
 
-	it('TODO:re1[repeated] Query with repeated path, different nested ids', async () => {
+	bench('TODO:re1[repeated] Query with repeated path, different nested ids', async () => {
 		const res = await ctx.query(
 			{
 				$entity: 'Space',
@@ -1629,7 +1629,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		});
 	});
 
-	it('TODO:re2[repeated] Query with repeated path, different nested patterns', async () => {
+	bench('TODO:re2[repeated] Query with repeated path, different nested patterns', async () => {
 		const res = await ctx.query(
 			{
 				$entity: 'Space',
@@ -1652,7 +1652,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		});
 	});
 
-	it('xf1[excludedFields] Testing excluded fields', async () => {
+	bench('xf1[excludedFields] Testing excluded fields', async () => {
 		let godUser = {
 			$entity: 'God',
 			id: 'squarepusher',
@@ -1690,7 +1690,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		});
 	});
 
-	it('xf2[excludedFields, deep] - deep nested', async () => {
+	bench('xf2[excludedFields, deep] - deep nested', async () => {
 		const query = {
 			$entity: 'Space',
 			$id: 'space-2',
@@ -1752,7 +1752,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		expect(deepSort(resWithoutMetadata, 'id')).toEqual(deepRemoveMetaData(expectedRes));
 	});
 
-	it('xf3[excludedFields, deep] - Exclude virtual field', async () => {
+	bench('xf3[excludedFields, deep] - Exclude virtual field', async () => {
 		const query = {
 			$entity: 'User',
 			$id: 'user2',
@@ -1790,7 +1790,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		expect(deepSort(resWithoutMetadata, 'id')).toEqual(deepRemoveMetaData(expectedRes));
 	});
 
-	it('vi1[virtual, attribute] Virtual DB field', async () => {
+	bench('vi1[virtual, attribute] Virtual DB field', async () => {
 		//This works with TypeDB rules
 		const res = await ctx.query({ $entity: 'Account', $fields: ['id', 'isSecureProvider'] }, { noMetadata: true });
 
@@ -1818,7 +1818,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		]);
 	});
 
-	it('vi2[virtual, edge] Virtual DB edge field', async () => {
+	bench('vi2[virtual, edge] Virtual DB edge field', async () => {
 		//This works with TypeDB rules
 		const res = await ctx.query({ $entity: 'Hook' }, { noMetadata: true });
 
@@ -1851,7 +1851,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		]);
 	});
 
-	it('co1[computed] Virtual computed field', async () => {
+	bench('co1[computed] Virtual computed field', async () => {
 		const res = await ctx.query(
 			{ $entity: 'Color', $id: ['blue', 'yellow'], $fields: ['id', 'isBlue'] },
 			{ noMetadata: true },
@@ -1869,7 +1869,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		]);
 	});
 
-	it('co2[computed] Computed virtual field depending on edge id', async () => {
+	bench('co2[computed] Computed virtual field depending on edge id', async () => {
 		const res = await ctx.query(
 			{ $entity: 'Color', $id: ['blue', 'yellow'], $fields: ['id', 'user-tags', 'totalUserTags'] },
 			{ noMetadata: true },
@@ -1889,7 +1889,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		]);
 	});
 
-	it('TODO:co3[computed], Computed virtual field depending on edge id, missing dependencies', async () => {
+	bench('TODO:co3[computed], Computed virtual field depending on edge id, missing dependencies', async () => {
 		const res = await ctx.query(
 			{ $entity: 'Color', $id: ['blue', 'yellow'], $fields: ['id', 'totalUserTags'] },
 			{ noMetadata: true },
@@ -1908,7 +1908,7 @@ export const testQuery = createTest('Query', (ctx) => {
 	});
 
 	/*
-  it('[entity,nested, filter] - $filter on children property', async () => {
+  bench('[entity,nested, filter] - $filter on children property', async () => {
     expect(bormClient).toBeDefined();
     const res = await bormClient.query({
       $entity: 'User',
@@ -1922,7 +1922,7 @@ export const testQuery = createTest('Query', (ctx) => {
       name: 'Antoine',
     });
   });
-  it('[entity,nested,filter] - Simplified filter', async () => {
+  bench('[entity,nested,filter] - Simplified filter', async () => {
     expect(bormClient).toBeDefined();
     const res = await bormClient.query({
       $entity: 'User',
@@ -1937,7 +1937,7 @@ export const testQuery = createTest('Query', (ctx) => {
       },
     ]);
   });
-  it('[entity,array,includes] - filter by field of cardinality many, type text: includes one ', async () => {
+  bench('[entity,array,includes] - filter by field of cardinality many, type text: includes one ', async () => {
     expect(bormClient).toBeDefined();
     const res = await bormClient.query({
       $entity: 'post',
@@ -1953,7 +1953,7 @@ export const testQuery = createTest('Query', (ctx) => {
       { $entity: 'post', $id: 'post2', id: 'post2' },
     ]);
   });
-  it('[entity,array,includesAll] - filter by field of cardinality many, type text: includes all ', async () => {
+  bench('[entity,array,includesAll] - filter by field of cardinality many, type text: includes all ', async () => {
     expect(bormClient).toBeDefined();
     const res = await bormClient.query({
       $entity: 'post',
@@ -1968,7 +1968,7 @@ export const testQuery = createTest('Query', (ctx) => {
       { $entity: 'post', $id: 'post3', id: 'post3' },
     ]);
   });
-  it('[entity,array,includesAny] filter by field of cardinality many, type text: includes any ', async () => {
+  bench('[entity,array,includesAny] filter by field of cardinality many, type text: includes any ', async () => {
     expect(bormClient).toBeDefined();
     const res = await bormClient.query({
       $entity: 'post',
@@ -1984,7 +1984,7 @@ export const testQuery = createTest('Query', (ctx) => {
       { $entity: 'post', $id: 'post3', id: 'post3' },
     ]);
   });
-  it('[entity,includesAny,error] using array filter includesAny on cardinality=ONE error', async () => {
+  bench('[entity,includesAny,error] using array filter includesAny on cardinality=ONE error', async () => {
     expect(bormClient).toBeDefined();
     const res = await bormClient.query({
       $entity: 'User',
@@ -1992,7 +1992,7 @@ export const testQuery = createTest('Query', (ctx) => {
     });
     expect(res).toThrow(TypeError);
   });
-  it('[entity,includesAll, error] using array filter includesAll on cardinality=ONE error', async () => {
+  bench('[entity,includesAll, error] using array filter includesAll on cardinality=ONE error', async () => {
     expect(bormClient).toBeDefined();
     const res = await bormClient.query({
       $entity: 'User',
@@ -2001,7 +2001,7 @@ export const testQuery = createTest('Query', (ctx) => {
     expect(res).toThrow(TypeError);
   });
   // OPERATORS: NOT
-  it('[entity,filter,not] - filter by field', async () => {
+  bench('[entity,filter,not] - filter by field', async () => {
     expect(bormClient).toBeDefined();
     const res = await bormClient.query({
       $entity: 'User',
@@ -2016,7 +2016,7 @@ export const testQuery = createTest('Query', (ctx) => {
       { $entity: 'User', $id: 'user2', id: 'user3' },
     ]);
   });
-  it('[entity,filter,not,array,includes] filter item cardinality many', async () => {
+  bench('[entity,filter,not,array,includes] filter item cardinality many', async () => {
     expect(bormClient).toBeDefined();
     const res = await bormClient.query({
       $entity: 'post',
@@ -2027,7 +2027,7 @@ export const testQuery = createTest('Query', (ctx) => {
   });
   // OPERATORS: OR
   // typeDB: https://docs.vaticle.com/docs/query/match-clause#disjunction-of-patterns. When is the same
-  it('[entity,OR] or filter two different fields', async () => {
+  bench('[entity,OR] or filter two different fields', async () => {
     expect(bormClient).toBeDefined();
     const res = await bormClient.query({
       $entity: 'User',
@@ -2046,7 +2046,7 @@ export const testQuery = createTest('Query', (ctx) => {
 
 	// NESTED
 
-	it('a1[$as] - as for attributes and roles and links', async () => {
+	bench('a1[$as] - as for attributes and roles and links', async () => {
 		const expectedRes = {
 			'email_as': 'antoine@test.com',
 			'id': 'user1',
@@ -2097,7 +2097,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		expect(deepSort(res, 'id')).toEqual(expectedRes);
 	});
 
-	it('bq1[batched query] - as for attributes and roles and links', async () => {
+	bench('bq1[batched query] - as for attributes and roles and links', async () => {
 		const expectedRes = [
 			{
 				id: 'user1',
@@ -2127,7 +2127,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		expect(res).toEqual(expectedRes);
 	});
 
-	it('j1[json] Query a thing with a JSON attribute', async () => {
+	bench('j1[json] Query a thing with a JSON attribute', async () => {
 		const entity = await ctx.query({
 			$entity: 'Account',
 			$id: 'account1-1',
@@ -2138,7 +2138,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		});
 	});
 
-	it('j2[json] Query a thing with an empty JSON attribute', async () => {
+	bench('j2[json] Query a thing with an empty JSON attribute', async () => {
 		const entity = await ctx.query({
 			$entity: 'Account',
 			$id: 'account1-2',
@@ -2147,7 +2147,7 @@ export const testQuery = createTest('Query', (ctx) => {
 		expect((entity as any).profile).toBeUndefined();
 	});
 
-	it('TODO:bq2[batched query with $as] - as for attributes and roles and links', async () => {
+	bench('TODO:bq2[batched query with $as] - as for attributes and roles and links', async () => {
 		const expectedRes = {
 			users: {
 				id: 'user1',