+++ changelog.md
@@ -2,6 +2,12 @@
 
 📝 following beta format X.Y.Z where Y = breaking change and Z = feature and fix. Later => FAIL.FEATURE.FIX
 
+## 0.10.13(2024-05-24)
+
+- Test: Added benchmark
+- Feat: SurrealDB queries
+-> $filter map to dbPath is now done only for TQL
+
 ## 0.10.12(2024-05-23)
 
 - Feat(WIP): SurrealDB fix empty arrays and tableName sanitizing 37✅/25❌